fix: combine useChatConversation hook into the context
